refactor: mv priceOracleKit
